make client partial/move files
