+++ AtCoder/Client/AtCoderClient.cs
@@ -6,9 +6,9 @@ using System.Net.Http;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
-namespace AtCoder.Web
+namespace AtCoder.Client
 {
-    public class AtCoderClient
+    public partial class AtCoderClient
     {
         const string revelSessionKey = "REVEL_SESSION";
         static readonly Uri AtCoderDomain = new Uri("https://atcoder.jp");
@@ -67,26 +67,6 @@ namespace AtCoder.Web
                 return null;
             }
         }
-
-        private async Task<T> GetParsedJsonResponseAsync<T>(string path)
-        {
-            var response = await Client.GetAsync(path);
-            if (response.StatusCode == HttpStatusCode.NotFound) return default;
-            if (!response.IsSuccessStatusCode) 
-                throw new WebException($"API returns {response.StatusCode}", WebExceptionStatus.ProtocolError);
-            string json = await response.Content.ReadAsStringAsync();
-            var result = JsonConvert.DeserializeObject<T>(json);
-            return result;
-        }
-
-        public Task<CompetitionResult[]> GetCompetitionHistoryAsync(string userScreenName)
-            => GetParsedJsonResponseAsync<CompetitionResult[]>($"/users/{Uri.EscapeDataString(userScreenName)}/history/json");
-    
-        public Task<Standings> GetStandingsAsync(string contestScreenName)
-            => GetParsedJsonResponseAsync<Standings>($"/contests/{Uri.EscapeDataString(contestScreenName)}/standings/json");
-
-        public Task<CompetitionResult[]> GetReslutsAsync(string contestScreenName)
-            => GetParsedJsonResponseAsync<CompetitionResult[]>($"/contests/{Uri.EscapeDataString(contestScreenName)}/results/json");
 
     }
 }
